FIFO is now an adapter configuration value
Update issue 21
The FIFO file to use to control MPlayer is set in the adapters configuration
file now. As default, the file ~/.cache/remuco/mplayer/fifo is used.
In a next step the MPlayer configuration could be adjusted automatically to
point to that FIFO. However, this should only be done if there is no such
option yet in the MPlayer config - otherwise users may get confused what
happened to their manually configured FIFO.