Add regex objects to match mplayer status infos
Update issue 21
Adding regexes for audio files (not tested) and video files (tested with AVIs,
but should be the same for every video file).
The groups are named, so you can get the matched information with something
like
	matchobj = self.video_regex.match(status_str)
	print matchobj.group('audiopos')

Client information retrieval is *still incomplete*. Ideally, __inputevent()
will modify a class member so that poll()'s next call will send the most
recent information to the client.