Exlclude JSR-82 files when JSR-82 lib is not set
The client build process now automatically excludes JSR-82 related
classes when the corresponding property in file 'build.properties' is
removed/commented.

Update issue 39
Status: Fixed
Labels: +Bluetooth
To build a client which works on devices using JBed, simply comment the
line 'jsr82.jar=...' in client/build.properties and then build the
client (see ClientCustomization page in the wiki). The resulting client
does not reference any JSR-82 classes - these references caused the
client to fail on JBed devices. Of couse, Bluetooth is not available
then, but this is a limitation of JBed.