Do not auto connect after a scan
Rationale is that this is consistent to the behavior when adding a WiFi
connection or when adding a manual BT connection.

Additionally the Device shown before doing a devices screen is not
attached to the scanning screen as it is not used afterwards.

New issue
Summary: Service options in BT screen are ignored when adding new device
Type: Bug
Status: Accepted
Owner: obensonne
Labels: Usability
When adding a new BT connection the service search options configured in
the BT screen are ignored.

The idea to fix this is to initially hide all options except
scan/address.