Add config option to turn on Jump-to option for MPRIS players
Currently Jump-to is disabled for MPRIS players since they are ambiguous
if the player uses dynamic playlists. However if dynamic playlists are
not being used this config option allows the user to turn on the Jump-to
option.

User will be able to enable the jump-to option for MPRIS players by setting
mpris-jump=1 in the respective player's config file.