Net: add power saving mechanism
Update issue 82
When the JavaME client receives a pause signal, it stops pinging the
server and sends a SLEEP message to the server. The server then stops
to send update messages (except BYE) to the client. Later, if a resume
signal arrives in the client, it re-enables pinging and sends a WAKEUP
message to the server which then restarts to send player information to
the client.

When pause and resume signals arrive in the client is completely up to
a phone's Java-VM. Usually it works for incoming phone calls or when the
client app get's 'minimalized' manually. Don't know if it works for
screensavers.

@yaroslav.khomyak:
Could you please test this? See
http://code.google.com/p/remuco/wiki/BleedingEdge for some instructions.