Base: simplify file browser configuration
The config options to en/disable the file browser feature and the option
to decide the usage of auto-detected mimetype dependent dirs.
Both options are now included by the `file-browser-root-dirs` option. If
this option is set to nothing, the file browser is disabled. If it
contains the keyword `auto`, mimetype-related dirs are included.