Doc: make README the one and only place for documentation
This fixes the problem of writing and maintaining docs online and
offline and for multiple versions. Now each version definitely has the
documentation contained in its package. At the same time the README of
the most recent version is used to generate the GettingStarted page in
the project wiki (check out the wiki repo to see how that works).

(Yes, some docs, e.g. the FAQ, are still wiki-only.)