Banshee: adapt shuffle control to Banshee 1.6
Update issue 126
Status: Fixed

When a client activates the shuffle mode, the `song` mode is used as
fedault. Other values (`artist`, `rating`, ..) can be set in Remuco's
configuration file using the option `x-shuffle` in the Bansee section.

Earlier versions of Banshee are not supported anymore (they work, but
without shuffle control and display).